Update the datatype for the maximum size of MaxBinarySize to use the Int32.Parse, not the Int16.Parse which is not a big enough size for sending attachments/documents like PDFs which can often be megabytes in size. (Int16.MaxSize = 32767, not really enough bytes... 2G should be enough...)

+++ Spark.Core/Settings.cs
@@ -42,13 +42,13 @@ namespace Spark.Config
             {
                 try
                 {
-                    int max = Convert.ToInt16(AppSettings.Get("MaxBinarySize"));
-                    if (max == 0) max = Int16.MaxValue;
+                    int max = Convert.ToInt32(AppSettings.Get("MaxBinarySize"));
+                    if (max == 0) max = Int32.MaxValue;
                     return max;
                 }
                 catch
                 {
-                    return Int16.MaxValue;
+                    return Int32.MaxValue;
                 }
             }
         }
